backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-Migrate
whisper-openai
